refactor(core): decompose Fact kernel internals
