+++ framework/core/architecture/TARGETS.cmake
@@ -58,7 +58,12 @@ set(KUNGFU_STORAGE_SERVICES_SOURCE_FILES
   "${PROJECT_SOURCE_DIR}/src/runtime/storage/episode_admission.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/storage/episode_manifest_projection.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/storage/episode_repair.cpp"
+  "${PROJECT_SOURCE_DIR}/src/runtime/storage/fact_commit.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/storage/fact_kernel.cpp"
+  "${PROJECT_SOURCE_DIR}/src/runtime/storage/fact_portability.cpp"
+  "${PROJECT_SOURCE_DIR}/src/runtime/storage/fact_protocol.cpp"
+  "${PROJECT_SOURCE_DIR}/src/runtime/storage/fact_query.cpp"
+  "${PROJECT_SOURCE_DIR}/src/runtime/storage/fact_state.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/storage/json_compat.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/storage/maintenance_service.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/storage/manifest_catalog_projection.cpp"
