+++ package.json
@@ -85,6 +85,7 @@
     "test:fact-kernel-dogfood": "node scripts/require-shifu.mjs test:fact-kernel-dogfood && node scripts/run-fact-kernel-dogfood-tests.mjs",
     "test:fact-kernel-native": "node scripts/require-shifu.mjs test:fact-kernel-native && PYTHONPATH=framework/core/src/python:framework/core/build/Release uv run --project framework/core --frozen pytest -q framework/core/tests/python/test_agent_work_profile_native.py framework/core/tests/python/test_fact_kernel_characterization.py",
     "check:fact-root-canonical": "node scripts/require-shifu.mjs check:fact-root-canonical && node --test scripts/check-fact-root-canonical.test.mjs",
+    "check:fact-kernel-boundary": "node scripts/require-shifu.mjs check:fact-kernel-boundary && node --test scripts/check-fact-kernel-boundary.test.mjs",
     "docs:check": "node scripts/require-shifu.mjs docs:check && node scripts/run-docs-check.mjs",
     "docs:check:readonly": "node scripts/require-shifu.mjs docs:check:readonly && node scripts/run-docs-readonly.mjs",
     "docs:prose": "node scripts/require-shifu.mjs docs:prose && node scripts/run-docs-prose-check.mjs",
